+++ test/keys.js
@@ -1,0 +1,55 @@
+module.exports.mockKeys = [
+	"person",
+	"phoneNumber",
+	"address"
+];
+
+module.exports.loremKeys = [
+	"sentence",
+	"paragraph",
+	"text"
+];
+
+module.exports.personKeys = [
+	"name",
+	"middlename",
+	"initials",
+	"middleInitial",
+	"lastname",
+	"sex",
+	"height",
+	"age",
+	"academicTitle",
+	"nationality",
+	"email",
+	"title",
+	"suffixTitle"
+];
+
+module.exports.usAddressKeys = [
+	"city",
+	"state",
+	"stateAbbrev",
+	"zipCode",
+	"countryAbbrev",
+	"country"
+];
+
+module.exports.usNumberKeys = [
+	"country",
+	"countryCode",
+	"cityCode",
+	"subNumber",
+	"number",
+	"withAreaCode"
+];
+
+module.exports.randomKeys = [
+	"number",
+	"title",
+	"imageUrl",
+	"imageBlob"
+];
+
+module.exports.randomImageUrlKeys = ["data", "info"];
+module.exports.randomImageBlobKeys = ["blob", "info"];